docs: clarify responsibilities of pairing submodules
